% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{bar_plot}
\alias{bar_plot}
\title{Create Publications Bar Charts (i.e., Publications by Tumor Site, Author, or Journal)}
\usage{
bar_plot(x, years = publications$year, display = NULL, x_label = x)
}
\arguments{
\item{x}{String representing data to plot on x axis}

\item{years}{Data years to include in analysis}

\item{display}{Number of bars to show on plot}

\item{x_label}{String for x-axis label}
}
\value{
ggplot2 bar plot
}
\description{
Create Publications Bar Charts (i.e., Publications by Tumor Site, Author, or Journal)
}
\examples{
\dontrun{
bar_plot("sites")
bar_plot("sites", years = c(2008:2015), display = 15)
}
}
