% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping_functions.R
\name{create_std_site_lookup}
\alias{create_std_site_lookup}
\title{Create lookup from publication site to SEER and CDC site}
\usage{
create_std_site_lookup(publications, seer, cdc)
}
\arguments{
\item{publications}{\code{data.table} publication data from running get_publication_data}

\item{seer}{\code{data.table} seer data from running get_seer_data}

\item{cdc}{\code{data.table} cdc data from running get_cdc_data}
}
\value{
\code{\link{data.table}}
}
\description{
Used to make sure a match between the SEER or CDC tumor site data and the publication data can be found.
}
