% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.arg_error}
\alias{.arg_error}
\title{Create Error Message - Column Not Found in Dataset}
\usage{
.arg_error(x, dataset)
}
\arguments{
\item{x}{String column name to check for in dataset}

\item{dataset}{dataset to search}
}
\value{
Appropriate error
}
\description{
Create Error Message - Column Not Found in Dataset
}
\examples{
\dontrun{}
}
