% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{.format_plot}
\alias{.format_plot}
\title{Format ggplot2 Plots With Minimalist Theme, Title, and Labels}
\usage{
.format_plot(plot, x_label, y_label = "Publications", title = "")
}
\arguments{
\item{plot}{ggplot2 plot to format}

\item{x_label}{String for x-axis label}

\item{y_label}{String for y-axis label}

\item{title}{String for plot title}
}
\value{
Formatted ggplot2 plot with labels
}
\description{
Format ggplot2 Plots With Minimalist Theme, Title, and Labels
}
\examples{
\dontrun{}
}
