% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{.format_table}
\alias{.format_table}
\title{Format SEER Tables Using ReporteRs FlexTable}
\usage{
.format_table(dt, title = "", footnote = "", less_than = NULL)
}
\arguments{
\item{dt}{data.table to use as FlexTable data}

\item{title}{String to use as table title}

\item{footnote}{String to use as table footnote}

\item{less_than}{numeric value to use as threshold for cell colors (cells with values less than this parameter are colored orange) -
if NULL, table has banded rows}
}
\value{
ReporteRs package FlexTable
}
\description{
Format SEER Tables Using ReporteRs FlexTable
}
\examples{
\dontrun{}
}
