% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.seer_label}
\alias{.seer_label}
\title{Format Seer Labels}
\usage{
.seer_label(x, sites, metadata, measure = tools::toTitleCase(x))
}
\value{
String in the form [SEER Data] per [SEER Data Scale] ([SEER Data Year]) - e.g., Incidence per 1,000 (2016)
}
\description{
Format Seer Labels
}
\examples{
\dontrun{}
}
