% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporters_functions.R
\name{.underrepresented}
\alias{.underrepresented}
\title{Determine Underrepresented Years in Publications Data}
\usage{
.underrepresented(seer_pub_data)
}
\value{
String containing underrepresented data years, or NULL, if no years are underrepresented
}
\description{
Assumes that publication data for any given year is fully represented 3 months after the year ends
Calculates if data is underrepresented using data access date and latest year in data
}
\examples{
\dontrun{}
}
