% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_figures.R
\name{generate_figures}
\alias{generate_figures}
\title{Generate Full Report: Overview of All SEER-Medicare Publications}
\usage{
generate_figures(seer_pub_data, outputPath = ".")
}
\arguments{
\item{seer_pub_data}{Result of running scrape_all_data.  Should have class "scrape_all_data"}

\item{outputPath}{Path to directory where report will be generated. Default is working directory}

\item{html}{Logical value - if TRUE, will generate html file; if FALSE, will generate Microsoft Word file}
}
\value{
doc Object of class \code{\link{docx}} or \code{\link{bsdoc}} containing SEER-Medicare report
}
\description{
Generate Full Report: Overview of All SEER-Medicare Publications
}
