% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{Generate Full Report: Overview of All SEER-Medicare Publications}
\usage{
generate_report(html = F)
}
\arguments{
\item{html}{Logical value - if TRUE, will generate html file; if FALSE, will generate Microsoft Word file}
}
\value{
doc Object of class \code{\link{docx}} or \code{\link{bsdoc}} containing SEER-Medicare report
}
\description{
Generate Full Report: Overview of All SEER-Medicare Publications
}
\examples{
\dontrun{
generate_report()
}
}
