% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping_functions.R
\name{get_cdc_data}
\alias{get_cdc_data}
\title{Get CDC United States Cancer Statistics Data: All Data Combined (Mortality and Incidence Rates)}
\usage{

  get_cdc_data(pathToZip = "http://www.cdc.gov/cancer/npcr/uscs/USCS_1999_2014_ASCII.zip")
}
\arguments{
\item{pathToZip}{\code{character} Path to zip file on cdc website}
}
\value{
\code{\link{data.table}}
}
\description{
Downloads and unzips pathToZip file from CDC website returns data.table version of BYAGE.TXT file which contains mortality and incidence rates by age group, race, sex, site, and year
}
\details{
see https://www.cdc.gov/cancer/npcr/uscs/download_data.htm which has links to all historical data
}
