% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_metadata}
\alias{get_metadata}
\title{Get Metadata Field}
\usage{
get_metadata(x)
}
\arguments{
\item{x}{String representing data to return}
}
\value{
Parameter metadata field
}
\description{
Get Metadata Field
}
\examples{
\dontrun{
get_metadata("access_date")
}
}
