% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping_functions.R
\name{get_publication_data}
\alias{get_publication_data}
\title{Get SEER Medicare based Publications data}
\usage{

  get_publication_data(nci_search_url = "http://healthcaredelivery.cancer.gov/cgi-bin-pubsearch/pubsearch/index.pl?simple_search=&searchOpt=and&initiative=SEERM",
  table_results_css_selector = "table.pubsearch-results")
}
\arguments{
\item{nci_search_url}{\code{character} URL that searches NCI publications for all SEER Medicare publications}

\item{table_results_css_selector}{\code{character} CSS selector to select main result table}
}
\value{
\code{\link{data.table}}
}
\description{
Searches https://healthcaredelivery.cancer.gov/publications/ for all SEER-Medicare Publications (Data Source/Project/Initiative = "SEER-Medicare Linked Database) and returns data.table with information about each publication
}
