% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_scale}
\alias{get_scale}
\title{Create Data Scales (used for incidence, prevalence, and deaths data)}
\usage{
get_scale(x, limit = 500)
}
\arguments{
\item{x}{Data column to use to calculate scale - should be "incidence", "prevalence", or "deaths"}

\item{limit}{Numeric limit - scales will be created to keep max values under this limit}
}
\value{
Base 10 scale calculated from max value and target limit (to use as denominator)
}
\description{
Create Data Scales (used for incidence, prevalence, and deaths data)
}
\examples{
\dontrun{
get_scale(sites$incidence)
}
}
