% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping_functions.R
\name{get_seer_data}
\alias{get_seer_data}
\title{Get SEER Incidence, Prevalence, and Mortality Data}
\usage{
get_seer_data(seer_url = "http://seer.cancer.gov/statfacts/more.html",
  link_css_selector = ".alphaList.rpw a")
}
\arguments{
\item{seer_url}{\code{character} Url to Main SEER cancer fact site that contains links to each cancer}

\item{link_css_selector}{\code{character} CSS selector to select all links to each cancer site}
}
\value{
\code{\link{data.table}}
}
\description{
Loops through each link at http://seer.cancer.gov/statfacts/more.html, which contains more information about different cancer sites.
}
\details{
Returns data set with a column for site, year, x (A categorical variable defining what is in the value column. Can be "incidence", "prevalence", or "deaths"), value ( the number associated with the measure in x)
}
