% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporters_functions.R
\name{level}
\alias{level}
\title{Set Unordered List Level}
\usage{
level(n)
}
\arguments{
\item{n}{Integer representing unordered list level}
}
\value{
parProperties object containing unordered list level
}
\description{
Set Unordered List Level
}
\examples{
\dontrun{
ReporteRs::addParagraph("mytext", par.properties = level(1))
}
}
