% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{line_plot}
\alias{line_plot}
\title{Create Publications by Year Plots}
\usage{
line_plot(publications, by = NULL, by_label = by)
}
\arguments{
\item{by}{String representing variable to use to cut data - will use top 4 most frequent in publications data}

\item{by_label}{Label to use in chart title}
}
\value{
ggplot2 line plot
}
\description{
Create Publications by Year Plots
}
\examples{
\dontrun{
line_plot()
line_plot(by = "sites", by_label = "tumor site")
}
}
