% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporters_functions.R
\name{myAddPlot}
\alias{myAddPlot}
\title{Add Plot to Document suppressing warnings}
\usage{
myAddPlot(doc, plot)
}
\arguments{
\item{doc}{Object of class \code{\link{docx}} where plot should be added}

\item{plot}{plot to be added to the document}
}
\value{
doc Object of class \code{\link{docx}}
}
\description{
Add Plot to Document suppressing warnings
}
