% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporters_functions.R
\name{nline}
\alias{nline}
\title{Add Blank Line to Report}
\usage{
nline(doc)
}
\arguments{
\item{doc}{Object of class \code{\link{docx}} where line should be added}
}
\value{
Object of class \code{\link{docx}} with line added
}
\description{
Add Blank Line to Report
}
\examples{
\dontrun{}
}
