% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{palette}
\alias{palette}
\title{Get Color Palette for Graphics}
\usage{
palette()
}
\value{
Character vector of colors
}
\description{
Get Color Palette for Graphics
}
\examples{
\dontrun{
palette()
}
}
