% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping_functions.R
\name{parse_seer_cancer_fact_page}
\alias{parse_seer_cancer_fact_page}
\title{Parse SEER Cancer Fact Page to get Latest Incidence, Prevalence, and Mortality Data}
\usage{
parse_seer_cancer_fact_page(link)
}
\arguments{
\item{link}{\code{character} Url to site specific SEER cancer fact site}
}
\value{
\code{\link{data.table}}
}
\description{
Takes link to SEER site specific cancer stat fact page and returns data set with a column for site, year, x (A categorical variable defining what is in the value column. Can be "incidence", "prevalence", or "deaths"), value ( the number associated with the measure in x)
}
