% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{scale_seer}
\alias{scale_seer}
\title{Scale Publications to Incidence, Prevalence, and Deaths Data}
\usage{
scale_seer()
}
\value{
data.table
}
\description{
Scale Publications to Incidence, Prevalence, and Deaths Data
}
\examples{
\dontrun{
scale_seer()
}
}
