% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping_functions.R
\name{scrape_all_data}
\alias{scrape_all_data}
\title{Gets all data needed to generate SEER Publication report}
\usage{

  scrape_all_data(nci_url = "http://healthcaredelivery.cancer.gov/cgi-bin-pubsearch/pubsearch/index.pl?simple_search=&searchOpt=and&initiative=SEERM",
  cdc_cancer_stat_url = "http://www.cdc.gov/cancer/npcr/uscs/USCS_1999_2014_ASCII.zip",
  seer_cancer_stats_url = "http://seer.cancer.gov/statfacts/more.html")
}
\arguments{
\item{nci_url}{\code{data.table} publication data from running get_publication_data}

\item{cdc_cancer_stat_url}{\code{character} Path to zip file on cdc website. Base url is http://www.cdc.gov/cancer/npcr/uscs/ whic contains multiple years of CDC cancer  data}

\item{seer_cancer_stats_url}{\code{character} Url to Main SEER cancer fact site that contains links to each cancer}
}
\value{
\code{seermedicare} obj
}
\description{
Returns
}
