% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{seer_cols}
\alias{seer_cols}
\title{Get Column Names for Relevant SEER Data}
\usage{
seer_cols()
}
\value{
Character vector column names
}
\description{
Get Column Names for Relevant SEER Data
}
\examples{
\dontrun{
seer_cols()
}
}
