% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{seer_plot}
\alias{seer_plot}
\title{Create Publications and Incidence/Prevalence/Deaths by Year Plot}
\usage{
seer_plot(x, seer_pub_data)
}
\arguments{
\item{x}{Data column to plot - should be "incidence", "prevalence", or "deaths"}
}
\value{
Combination ggplot2 plot containing bar plot for parameter and line plot for total publications
}
\description{
Create Publications and Incidence/Prevalence/Deaths by Year Plot
}
\examples{
\dontrun{
seer_plot("incidence")
seer_plot("prevalence")
seer_plot("deaths")
}
}
