% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{seer_table}
\alias{seer_table}
\title{Create SEER Tables}
\usage{
seer_table(seer_pub_data, standardize = F, ...)
}
\arguments{
\item{standardize}{Logical value - if TRUE, standardizes publications ratios to median}

\item{...}{Further arguments passed to .format_table()}
}
\value{
ReporteRs package FlexTable containing total publications, % incident patients age 65 or older, publications to incidence ratios,
publications to prevalence ratios, and publications to deaths ratios for each tumor site
}
\description{
Create SEER Tables
}
\examples{
\dontrun{
seer_table()
seer_table(standardize = T)
}
}
