% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping_functions.R
\name{summarize_cdc_incidence_data_by_site}
\alias{summarize_cdc_incidence_data_by_site}
\title{Summarize CDC incidence data by standardized cancer site}
\usage{
summarize_cdc_incidence_data_by_site(cdc, cdc_year = "2010-2014")
}
\arguments{
\item{cdc}{\code{data.table} from running get_cdc_data and creating site_std variable which is the site standardized to publication site}

\item{cdc_year}{\code{character} The cdc data has results for each individual year and a summary of the last 4 years in the form (YYYY-YYYY). As new data comes in this variable will need to be udpated to match the newest 4 year summary.}
}
\value{
\code{\link{data.table}}
}
\description{
Summarize CDC incidence data by standardized cancer site
}
