% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping_functions.R
\name{summarize_seer_data_by_site}
\alias{summarize_seer_data_by_site}
\title{Summarize SEER data by standardized cancer site}
\usage{
summarize_seer_data_by_site(seer)
}
\arguments{
\item{seer}{\code{data.table} from running get_seer_data and creating site_std variable which is the site standardized to publication site}
}
\value{
\code{\link{data.table}} Summarized seer data by measure ("incidence", "prevalence", "deaths"), site, and year
}
\description{
Returns summarized counts for each measure
}
