% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping_functions.R
\name{summarize_site_data}
\alias{summarize_site_data}
\title{Summarize Site Data from Publication, CDC, and SEER}
\usage{
summarize_site_data(publications, seer, cdc, cdc_year = cdc[nchar(year) > 4,
  unique(year)])
}
\arguments{
\item{publications}{\code{data.table} publication data from running get_publication_data}

\item{seer}{\code{data.table} seer data from running get_seer_data}

\item{cdc}{\code{data.table} cdc data from running get_cdc_data}

\item{cdc_year}{\code{character} The cdc data has results for each individual year and a summary of the last 4 years in the form (YYYY-YYYY). As new data comes in this variable will need to be udpated to match the newest 4 year summary. Default gets the year from the cdc data that has more then 4 characters (all other years are single year YYYY)}
}
\value{
\code{\link{data.table}} Summarized seer data by measure ("incidence", "prevalence", "deaths"), site, and year
}
\description{
Summarize Site Data from Publication, CDC, and SEER
}
